Advanced_RAG/pyproject.toml
[project]
name = "advanced-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.1",
    "google-genai>=1.65.0",
    "langchain>=1.2.10",
    "langchain-text-splitters>=1.1.1",
    "nltk>=3.9.3",
    "numpy>=2.4.2",
    "pinecone>=8.1.0",
    "pypdf>=6.7.3",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "redis>=7.2.1",
    "tiktoken>=0.12.0",
    "uvicorn>=0.41.0",
]
